refactor(console): refactor console routes
