+++ packages/console/src/cloud/App.tsx
@@ -4,7 +4,6 @@ import Callback from '@cloud/pages/Callback';
 
 import * as styles from './App.module.scss';
 import Main from './pages/Main';
-import Onboard from './pages/Onboard';
 import { CloudRoute } from './types';
 
 const App = () => {
@@ -12,7 +11,6 @@ const App = () => {
     <BrowserRouter>
       <div className={styles.app}>
         <Routes>
-          <Route path={`/${CloudRoute.Onboard}/*`} element={<Onboard />} />
           <Route path={`/${CloudRoute.Callback}`} element={<Callback />} />
           <Route path={`/:tenantId/${CloudRoute.Callback}`} element={<Callback />} />
           <Route path="/*" element={<Main />} />
