+++ packages/console/src/containers/AppContent/components/UserInfoSkeleton/index.module.scss
