+++ packages/console/src/containers/ConsoleContent/Sidebar/components/Item/index.module.scss
