+++ packages/console/src/containers/ConsoleContent/Sidebar/components/Section/index.module.scss
