+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/BarGraph.tsx
