+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Bolt.tsx
