+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Box.tsx
