+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Cloud.tsx
