+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Connection.tsx
