+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Document.tsx
