+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Gear.tsx
