+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/List.tsx
