+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/UserProfile.tsx
