+++ packages/console/src/containers/ConsoleContent/Sidebar/icons/Web.tsx
