+++ packages/console/src/containers/ConsoleContent/Sidebar/index.module.scss
