+++ packages/console/src/containers/ConsoleContent/Sidebar/index.tsx
