+++ packages/console/src/containers/ConsoleContent/index.module.scss
@@ -1,0 +1,17 @@
+@use '@/scss/underscore' as _;
+
+.content {
+  flex-grow: 1;
+  display: flex;
+  overflow: hidden;
+}
+
+.main {
+  flex-grow: 1;
+  padding: 0 _.unit(2);
+  overflow-y: scroll;
+
+  > * {
+    @include _.main-content-width;
+  }
+}
